[TES3MP-deploy] Code cleaning around Terra handling

+++ TES3MP-deploy/install.sh
@@ -95,8 +95,7 @@ git clone https://github.com/TES3MP/openmw-tes3mp.git "$CODE"
 if [ $BUILD_OSG ]; then git clone https://github.com/openscenegraph/OpenSceneGraph.git "$DEPENDENCIES"/osg; fi
 if [ $BUILD_BULLET ]; then git clone https://github.com/bulletphysics/bullet3.git "$DEPENDENCIES"/bullet; fi
 git clone https://github.com/TES3MP/RakNet.git "$DEPENDENCIES"/raknet
-#git clone https://github.com/zdevito/terra.git "$DEPENDENCIES"/terra
-wget https://github.com/zdevito/terra/releases/download/release-2016-02-26/terra-Linux-x86_64-2fa8d0a.zip -O "$DEPENDENCIES"/terra.zip
+if [ $BUILD_TERRA ]; then git clone https://github.com/zdevito/terra.git "$DEPENDENCIES"/terra; else wget https://github.com/zdevito/terra/releases/download/release-2016-02-26/terra-Linux-x86_64-2fa8d0a.zip -O "$DEPENDENCIES"/terra.zip; fi
 git clone https://github.com/TES3MP/PluginExamples.git "$KEEPERS"/PluginExamples
 
 #COPY STATIC SERVER AND CLIENT CONFIGS
@@ -152,17 +151,18 @@ ln -s "$DEPENDENCIES"/raknet/include/RakNet "$DEPENDENCIES"/raknet/include/rakne
 
 cd "$BASE"
 
-##BUILD TERRA
-#echo -e "\n>> Building Terra"
-#cd $DEPENDENCIES/terra/
-#make -j$CORES
-
-#UNPACK AND PREPARE TERRA
-echo -e "\n>> Unpacking and preparing Terra"
-cd "$DEPENDENCIES"
-unzip terra.zip
-mv terra-* terra
-rm terra.zip
+#BUILD TERRA
+if [ $BUILD_TERRA ]; then
+    echo -e "\n>> Building Terra"
+    cd $DEPENDENCIES/terra/
+    make -j$CORES
+else
+    echo -e "\n>> Unpacking and preparing Terra"
+    cd "$DEPENDENCIES"
+    unzip terra.zip
+    mv terra-* terra
+    rm terra.zip
+fi
 
 cd "$BASE"
